+++ src/staking/rest_api/apis/mod.rs
@@ -15,5 +15,7 @@ pub mod eth_staking_api;
 pub use eth_staking_api::*;
 pub mod on_chain_yields_api;
 pub use on_chain_yields_api::*;
+pub mod soft_staking_api;
+pub use soft_staking_api::*;
 pub mod sol_staking_api;
 pub use sol_staking_api::*;