+++ src/wallet/rest_api/mod.rs
@@ -881,9 +881,6 @@ impl RestApi {
     ///
     /// Query User Wallet Balance
     ///
-    ///
-    /// * You need to open Permits Universal Transfer permission for the API Key which requests this endpoint.
-    ///
     /// Weight: 60
     ///
     /// # Arguments